fix: deleted some files, moved things around, need to fix ciruclar dependency
