+++ libs/trucking-kit/home/feature-home/src/index.ts
@@ -1,2 +1,3 @@
 export * from './lib/home.component';
 export * from './lib/home.routes';
+export * from './lib/companies/companies.component';
